pyproject.toml
[tool.poetry]
name = "books"
version = "0.1.0"
description = "A tool for making an inventory of a book collection"
authors = ["Ben Steinberg <bsteinberg@law.harvard.edu>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.0"
click = "^8.1.3"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
